Now UV & OV flags for LTC6812 & LTC6813 working
driverSWLTC6804ReadVoltageFlags();
driverSWLTC6804ReadStatusValues();

Those funtions can now read UV & OV flags above 12S which are located in aux groud D registers

undervoltage & overvoltage are uint32_t instead of uint16_t... for enabling more than 16 cells per IC...

+++ Drivers/SWDrivers/Inc/driverSWLTC6803.h
@@ -110,7 +110,7 @@ void driverSWLTC6803WriteConfig(driverLTC6803ConfigStructTypedef configStruct);
 bool driverSWLTC6803ReadConfigRegisters(uint8_t total_ic, uint8_t r_config[][7]);
 bool driverSWLTC6803ReadConfig(driverLTC6803ConfigStructTypedef *configStruct);
 bool driverSWLTC6803ReadFlagRegisters(uint8_t total_ic, uint8_t flagRegisters[][4]);
-bool driverSWLTC6803ReadVoltageFlags(uint16_t *underVoltageFlags, uint16_t *overVoltageFlags);
+bool driverSWLTC6803ReadVoltageFlags(uint32_t *underVoltageFlags, uint32_t *overVoltageFlags);
 void driverSWLTC6803EnableBalanceResistors(uint16_t balanceEnableMask);
 
 // Calculating CRC for communication
